Apply Double type on seekTo() and getCurrentTime()
- milliseconds 계산을 위해 int에서 double로 type 변경

+++ android/src/main/java/com/inprogress/reactnativeyoutube/YouTubeView.java
@@ -123,7 +123,7 @@ public class YouTubeView extends FrameLayout {
     public void didChangeToSeeking(int milliSeconds) {
         WritableMap event = Arguments.createMap();
         event.putString("state", "seeking");
-        event.putDouble("currentTime", milliSeconds / 1000);
+        event.putDouble("currentTime", (double)milliSeconds / 1000);
         event.putInt("target", getId());
         ReactContext reactContext = getReactContext();
         reactContext.getJSModule(RCTEventEmitter.class).receiveEvent(getId(), "state", event);
